backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "unified-multimodal-backend"
version = "0.1.0"
description = "Backend service for unified multimodal model management platform"
readme = "README.md"
authors = [{ name = "Unified Platform Team", email = "dev@unimodal.local" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
    "sqlmodel>=0.0.22,<0.0.23",
    "aiosqlite>=0.20.0,<0.21.0",
    "greenlet>=3.0.0,<4.0.0",
    "psycopg[binary]>=3.2.1,<4.0.0",
    "redis>=5.0.8,<6.0.0",
    "celery>=5.4.0,<6.0.0",
    "httpx>=0.27.0,<0.28.0",
    "pydantic-settings>=2.5.2,<3.0.0",
    "huggingface-hub>=0.24.0,<0.25.0",
    "psutil>=6.0.0,<7.0.0",
    "pynvml>=11.5.0,<12.0.0",
    "GPUtil>=1.4.0,<2.0.0",
    "loguru>=0.7.2,<0.8.0",
    "python-multipart>=0.0.9,<0.0.10",
    "py-cpuinfo>=9.0.0,<10.0.0",
    "openai-whisper>=20231117",
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.2,<9.0.0",
    "pytest-asyncio>=0.23.8,<0.24.0",
    "ruff>=0.6.0,<0.7.0",
    "httpx[cli]>=0.27.0,<0.28.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
